pyproject.toml
[tool.poetry]
name = "bot-folker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^20.6"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
mypy = "^1.6.1"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
